separated all data classes
