+++ Assets/GrandmaObjectCore/Agent/Data/AgentItemData.cs
@@ -1,0 +1,14 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using System;
+namespace Grandma
+{
+    [Serializable]
+    public class AgentItemData : GrandmaComponentData
+    {
+        [HideInInspector]
+        public string agentID;
+    }
+
+}
