+++ Assets/GrandmaObjectCore/Agent/Data/AgentItemData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 53cfa6eaeed9849a1bed3eadcaf59ea9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
