+++ Assets/GrandmaObjectCore/Interactable/DataClasses.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 153fe89e839f346d494952a9fc455b3d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
