+++ Assets/GrandmaObjectCore/Interactable/DataClasses/DoorData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9686597bf66d14e938576d2406c43ce2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
