+++ Assets/GrandmaObjectCore/Interactable/DataClasses/KeyData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 48ca7dccf26a24ebda1d3fdcc606af17
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
