+++ Assets/GrandmaObjectCore/Positionables/Controllers/DataClasses.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b7a05374c16f748ea977b462688e3812
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
