+++ Assets/GrandmaObjectCore/Positionables/Controllers/DataClasses/GroundMovement2DData.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace Grandma
+{
+    [CreateAssetMenu(menuName = "Core/MovementData/2D/GroundMovement2DData")]
+    public class GroundMovement2DData : GrandmaComponentData
+    {
+        public float speedScalar;
+        //public float jumpForce;
+    }
+}
