+++ Assets/GrandmaObjectCore/Positionables/Controllers/DataClasses/GroundMovement2DData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9097c995b82cf4146bccf29d025d3c3e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
