+++ Assets/GrandmaObjectCore/Positionables/Controllers/DataClasses/GroundMovementData.cs
@@ -1,0 +1,14 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+namespace Grandma
+{
+    [CreateAssetMenu(menuName = "Core/GroundMovementData")]
+    public class GroundMovementData : PositionableData
+    {
+        [SerializeField]
+        public float speedScalar;
+        [SerializeField]
+        public float jumpForce;
+    }
+}
