+++ Assets/GrandmaObjectCore/Positionables/Controllers/DataClasses/GroundMovementData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c649eb921b1e244929be41d58e5647c9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
