+++ Assets/GrandmaObjectCore/Positionables/Controllers/DataClasses/ZeroGMovementData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a97ac1e351355444c85d5f212b078bf9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
