+++ Assets/GrandmaObjectCore/Positionables/Controllers/GroundMovement2D.cs
@@ -4,13 +4,6 @@ using UnityEngine;
 
 namespace Grandma
 {
-    [CreateAssetMenu(menuName = "Core/MovementData/2D/GroundMovement2DData")]
-    public class GroundMovement2DData : GrandmaComponentData
-    {
-        public float speedScalar;
-        //public float jumpForce;
-    }
-
     public class GroundMovement2D : RBMove2D
     {
         [System.NonSerialized]
